move methods to web class
